+++ lib/overlay_layers.dart
@@ -8,12 +8,14 @@ export 'src/popup/popup_controller.dart' show PopupController, PopupDataContext;
 export 'src/popup/popup_widgets.dart'
     show PopupScaffold, AnimatedPopup, PositionedPopup, PopupPosition;
 
+// Modal exports
+export 'src/modal/modal_controller.dart' show ModalController, ModalDataContext;
+export 'src/modal/modal_widgets.dart'
+    show ModalScaffold, AnimatedModal, SlideUpModal;
+
 // Future exports will be added here:
 // Toast functionality
 // export 'src/toast/toast_controller.dart';
-
-// Modal functionality
-// export 'src/modal/modal_controller.dart';
 
 // Dialog functionality
 // export 'src/dialog/dialog_controller.dart';